+++ packages/internal-build-utils/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log - @typespec/internal-build-utils
 
+## 0.82.0
+
+No changes, version bump only.
+
 ## 0.81.0
 
 No changes, version bump only.